% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dicmo.R
\name{summary.dicmo}
\alias{summary.dicmo}
\title{S3 method for class 'dicmo'}
\usage{
\method{summary}{dicmo}(object, ...)
}
\arguments{
\item{object}{A 'dicmo' class object}

\item{...}{Other parameters that are not defined}
}
\description{
summary results for class 'dicmo'
}
\author{
X.PAN
}
