% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Xmnl.R
\name{summary.Xmnl}
\alias{summary.Xmnl}
\title{S3 method for class 'Xmnl'}
\usage{
\method{summary}{Xmnl}(object, ...)
}
\arguments{
\item{object}{A 'Xmnl' class object}

\item{...}{Other parameters that are not defined}
}
\description{
summary results for class 'Xmnl'
}
\author{
Xiaofeng Pan
}
