% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.explode.R
\name{X.explode}
\alias{X.explode}
\title{function to estimate exploded logit/probit model, including BWS case 3}
\usage{
X.explode(
  data,
  choice,
  alts,
  attrs,
  attr_coding = NULL,
  attr_level = NULL,
  interact = NULL,
  type = "logit",
  nest,
  nest_uni = TRUE,
  bw = FALSE,
  scale = FALSE,
  rank_alts = FALSE,
  avi = NULL,
  method = "BFGS",
  estimator = TRUE,
  param_fixed = NULL,
  param_start = NULL
)
}
\arguments{
\item{data}{A tibble, input data, wide format.}

\item{choice}{A vector of character, name of column indicating individuals'
choices from the most to the worst - the order matters! This argument is
useless and not necessary if the argument "rank_alts" is TRUE.}

\item{alts}{A vector of characters, names of all alternatives, including
the none-option if any.}

\item{attrs}{A list contains three slices, whose elements have to be non-
negative integers. The first one is a tibble, named "attrs_alts",
indicating the alternative-specific attributes (excluding ASCs); the
second one, named "asc", is a vector indicating the ASCs, and the third
one, named "context", is a tibble indicating the context variables
(includig individuals' socio-demographics). The column names in all
tibbles represent the name of attributes, or context variables. All have
same numbers of rows (elements) with the length of argument "alts". The
element in tibbles/vector indicates if the attribute, constant or
context variable is alternative-specific: "0" means the attribute,
constaant or context variable is not available for the alternative
(based on the sequence of alternative in the argument "alts"); except
"0", if some elements in one column have a same value, then the
corresponding alternatives have generic parameter in terms of this
attribute, constant or context variable.}

\item{attr_coding}{A vector of character, names of categorical attributes.
Default = NULL, which means all attributes are continuous.}

\item{attr_level}{A list of vector of character, representing the levels of
all categorical attributes. The first element of a vector is set as base
level, others are set as 1, 2, 3...in sequence. This argument is only
relevent when argument "attr_coding" is not NULL. Default = NULL.}

\item{interact}{A vector of character, name of attributes' interaction,
connected by "*". Default = NULL.}

\item{type}{A character, indicating which type of model is used, either
"logit" or "nl2". Default = "logit".}

\item{nest}{A list, indicating how many nests at all and which alternatives
are in the same nest.}

\item{nest_uni}{A logical, indicating if the inclusive values for the nests
are the same or not. Default = TRUE, means the inclusive values are the
same.}

\item{bw}{A logical, if TRUE, then BWS case 3 is estimated, otherwise the
complete rank is used. Default = FALSE.}

\item{scale}{A logical, if TRUE, alternative-specific scale parametes are
estimated, otherwise it is fixed to 1. Note only alternative-specific
scale parameters are allowed. Default = FASLE.}

\item{rank_alts}{A logical, if TRUE, the alternatives are ranked otherwise
the preferences are ranked. Default = FALSE.}

\item{avi}{A character, name of column indicating if an alternative is
available to individuals. Default = NULL, indicating all alternatives are
available to all respondents. Each alternative should have such a column,
for example, avi = "available" then the column's name for an alternative
(e.g. the alternative's name is "car") should be "available:car" or
"car:avilable". If this parameter is NULL, then those columns are not
necessary. If this parameter is not NULL, then in such as column, the
element should be 0 if the alternative is not available otherwise 1.}

\item{method}{A character, passed to the function maxLik() in "maxLik"
package. It indicates the method used in maximum likelihood estimation.
Default = "BFGS".}

\item{estimator}{A argument in the function maxLik() from "maxLik"
package. Here this argument could only be either TRUE or "BHHH"/"bhhh".
Default = TRUE.}

\item{param_fixed}{A vector of characters, passed to the function maxLik() in
"maxLik" package. It indicates which parameters are fixed. Default = NULL.}

\item{param_start}{A vector of numbers, passed to the function maxLik() in
"maxLik" package. It indicages the initial values of parameters.
Default = NULL.}
}
\description{
This function could estimate multinomial logit model along with
    converting the data from a wide format to a long format and code the
    categorical attributesallow allow. In detail, it allows to estimate
    interaction effects between attributes and alternative-specific
    parameters.
}
\author{
X.PAN
}
