% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.order.R
\name{X.order}
\alias{X.order}
\title{function to estimate ordered choice model}
\usage{
X.order(
  data,
  choice,
  rate,
  attrs,
  attr_coding = NULL,
  attr_level = NULL,
  interact = NULL,
  type = "ologit",
  method = "BFGS",
  estimator = TRUE,
  param_fixed = NULL,
  param_start = NULL
)
}
\arguments{
\item{data}{A tibble, input data.}

\item{choice}{A character, name of column indicating individuals' rate.}

\item{rate}{A vector of characters, the scale that used ot rate an
alternative.}

\item{attrs}{A list contains three slices, whose elements have to be non-
negative integers. The first one is a tibble, named "attrs_alts",
indicating the alternative-specific attributes (excluding ASCs); the
second one, named "asc", is a vector indicating the ASCs, and the third
one, named "context", is a tibble indicating the context variables
(includig individuals' socio-demographics). The column names in all
tibbles represent the name of attributes, or context variables. All have
same numbers of rows (elements) with the length of argument "alts". The
element in tibbles/vector indicates if the attribute, constant or
context variable is alternative-specific: "0" means the attribute,
constaant or context variable is not available for the alternative
(based on the sequence of alternative in the argument "alts"); except
"0", if some elements in one column have a same value, then the
corresponding alternatives have generic parameter in terms of this
attribute, constant or context variable.}

\item{attr_coding}{A vector of character, names of categorical attributes.
Default = NULL, which means all attributes are continuous.}

\item{attr_level}{A list of vector of character, representing the levels of
all categorical attributes. The first element of a vector is set as base
level, others are set as 1, 2, 3...in sequence. This argument is only
relevent when argument "attr_coding" is not NULL. Default = NULL.}

\item{interact}{A vector of character, name of attributes' interaction,
connected by "*". Default = NULL.}

\item{type}{A character, indicating which type of ordered model is used,
either "ologit" or "oprobit". Default = "ologit".}

\item{method}{A character, passed to the function maxLik() in "maxLik"
package. It indicates the method used in maximum likelihood estimation.
Default = "BFGS".}

\item{estimator}{A argument in the function maxLik() from "maxLik"
package. Here this argument could only be either TRUE or "BHHH"/"bhhh".
Default = TRUE.}

\item{param_fixed}{A vector of characters, passed to the function maxLik() in
"maxLik" package. It indicates which parameters are fixed. Default = NULL.}

\item{param_start}{A vector of numbers, passed to the function maxLik() in
"maxLik" package. It indicages the initial values of parameters.
Default = NULL.}
}
\description{
estimate ordered choice model
}
\author{
X.PAN
}
