% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L.wide2long.R
\name{L.wide2long}
\alias{L.wide2long}
\title{reshape the data from wide format to long format}
\usage{
L.wide2long(data, choice, alts, avi = NULL)
}
\arguments{
\item{data}{A tibble, input data.}

\item{choice}{A character, name of column that is individuals' choice.}

\item{alts}{A vector of characters, names of all alternatives, including the
none-option if any.}

\item{avi}{A character, name of column indicating if an alternative is
available to individuals. Default = NULL, indicating all alternatives are
available to all respondents. Each alternative should have such a column,
for example, avi = "available" then the column's name for an alternative
(e.g. the alternative's name is "car") should be "available:car" or
"car:avilable". If this parameter is NULL, then those columns are not
necessary. If this parameter is not NULL, then in such as column, the
element should be 0 if the alternative is not available otherwise 1.}
}
\description{
reshape the data from wide format to long format
}
\author{
Xiaofeng Pan
}
